fix: moves sumbission function do seperate api file, moves api url to env file, adds useFormStatus hook to control form status
