+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import { Form } from './components/Form'
+import { Form } from './components/Form/Form'
 
 function App() {
   return (
